Complete move of models to participants app
