+++ manage_breast_screening/participants/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from .models import Participant, ParticipantAddress
+from .models import Appointment, Participant, ParticipantAddress, ScreeningEpisode
 
 
 class AddressInline(admin.TabularInline):
@@ -11,4 +11,19 @@ class ParticipantAdmin(admin.ModelAdmin):
     inlines = [AddressInline]
 
 
+class AppointmentAdmin(admin.ModelAdmin):
+    list_display = [
+        "name",
+        "clinic_slot__starts_at",
+        "clinic_slot__duration_in_minutes",
+        "status",
+    ]
+
+    @admin.display()
+    def name(self, obj):
+        return obj.screening_episode.participant.full_name
+
+
 admin.site.register(Participant, ParticipantAdmin)
+admin.site.register(Appointment, AppointmentAdmin)
+admin.site.register(ScreeningEpisode)
