+++ manage_breast_screening/record_a_mammogram/tests/system/test_recording_a_mammogram.py
@@ -4,12 +4,12 @@ import pytest
 from django.urls import reverse
 from playwright.sync_api import expect
 
-from manage_breast_screening.clinics.tests.factories import (
+from manage_breast_screening.config.system_test_setup import SystemTestCase
+from manage_breast_screening.participants.tests.factories import (
     AppointmentFactory,
+    ParticipantFactory,
     ScreeningEpisodeFactory,
 )
-from manage_breast_screening.config.system_test_setup import SystemTestCase
-from manage_breast_screening.participants.tests.factories import ParticipantFactory
 
 
 class TestRecordingAMammogram(SystemTestCase):
